# print_agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Optica Print Agent API",
    "description": "API local para detectar impresoras Windows, configurar impresoras por tipo de documento y enviar trabajos de impresion POS o texto.",
    "contact": {
      "name": "Optica"
    },
    "license": {
      "name": "Uso interno"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Verificar estado del servicio",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/printers": {
      "get": {
        "tags": [
          "printers"
        ],
        "summary": "Listar impresoras detectadas",
        "operationId": "get_printers_printers_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrinterInfo"
                  },
                  "title": "Response Get Printers Printers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/printers/test": {
      "post": {
        "tags": [
          "printers"
        ],
        "summary": "Enviar impresion de prueba",
        "operationId": "test_printer_printers_test_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestPrintRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Test Printer Printers Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/config": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "Consultar configuracion actual",
        "operationId": "get_config_config_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/config/printers": {
      "put": {
        "tags": [
          "config"
        ],
        "summary": "Guardar impresora por tipo de documento",
        "operationId": "set_default_printer_config_printers_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrinterSelectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Set Default Printer Config Printers Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/print": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Crear e imprimir un trabajo",
        "operationId": "enqueue_print_job_jobs_print_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrintJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Listar trabajos de impresion",
        "operationId": "get_jobs_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrintJobResponse"
                  },
                  "title": "Response Get Jobs Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Consultar un trabajo de impresion",
        "operationId": "get_job_detail_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentConfigResponse": {
        "properties": {
          "api_token_configured": {
            "type": "boolean",
            "title": "Api Token Configured"
          },
          "host": {
            "type": "string",
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "printers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Printers"
          }
        },
        "type": "object",
        "required": [
          "api_token_configured",
          "host",
          "port",
          "printers"
        ],
        "title": "AgentConfigResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PrintJobRequest": {
        "properties": {
          "document_type": {
            "type": "string",
            "enum": [
              "constancia",
              "remision",
              "factura",
              "recibo",
              "abono",
              "pedido"
            ],
            "title": "Document Type"
          },
          "format": {
            "type": "string",
            "enum": [
              "pos",
              "text",
              "pdf"
            ],
            "title": "Format",
            "default": "pos"
          },
          "printer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Printer Name"
          },
          "copies": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Copies",
            "default": 1
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "document_type"
        ],
        "title": "PrintJobRequest"
      },
      "PrintJobResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "printer_name": {
            "type": "string",
            "title": "Printer Name"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "copies": {
            "type": "integer",
            "title": "Copies"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "generated_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated File"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "printer_name",
          "document_type",
          "format",
          "copies",
          "created_at"
        ],
        "title": "PrintJobResponse"
      },
      "PrinterInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "port": {
            "type": "string",
            "title": "Port",
            "default": ""
          },
          "driver": {
            "type": "string",
            "title": "Driver",
            "default": ""
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PrinterInfo"
      },
      "PrinterSelectionRequest": {
        "properties": {
          "document_type": {
            "type": "string",
            "title": "Document Type",
            "examples": [
              "constancia"
            ]
          },
          "printer_name": {
            "type": "string",
            "title": "Printer Name",
            "examples": [
              "EPSON TM-T20II Receipt"
            ]
          }
        },
        "type": "object",
        "required": [
          "document_type",
          "printer_name"
        ],
        "title": "PrinterSelectionRequest"
      },
      "TestPrintRequest": {
        "properties": {
          "printer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Printer Name"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Prueba de impresion desde Optica Print Agent"
          }
        },
        "type": "object",
        "title": "TestPrintRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "health",
      "description": "Verificacion del estado del servicio."
    },
    {
      "name": "printers",
      "description": "Deteccion y prueba de impresoras."
    },
    {
      "name": "config",
      "description": "Configuracion persistente del agente."
    },
    {
      "name": "jobs",
      "description": "Creacion y consulta de trabajos de impresion."
    }
  ]
}